Separate the view for clearer code
