+++ aidants_connect_web/views/__init__.py
@@ -1,0 +1,2 @@
+from .service import *
+from .FC_as_FS import *
